pyproject.toml
[tool.poetry]
name = "icloudbridge"
version = "0.1.0"
description = "Simplified API-based synchronization for Apple Notes & Reminders to NextCloud, CalDAV, and local folders"
authors = ["Keith Vassallo <keith@vassallo.cloud>"]
readme = "README.md"
license = "GPL-3.0-or-later"
homepage = "https://taskbridge.app/"
repository = "https://github.com/keithvassallomt/icloudbridge"
keywords = ["sync", "note sync", "reminder sync", "cloud notes", "cloud reminders", "apple", "icloud"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Environment :: MacOS X",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Topic :: Utilities"
]
packages = [{include = "icloudbridge"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
typer = {extras = ["all"], version = "^0.20.0"}
caldav = "^1.3.0"
aiofiles = "^23.2.0"
aiosqlite = "^0.19.0"
httpx = "^0.26.0"
markdown-it-py = "^3.0.0"
html-to-markdown = "^2.5.0"
keyring = "^24.3.0"
pyobjc-framework-EventKit = "^12.0"
pyobjc-framework-Cocoa = "^12.0"
tomli-w = "^1.0.0"
apscheduler = "^3.10.4"
python-multipart = "^0.0.6"
websockets = "^12.0"
email-validator = "^2.1.0"
argon2-cffi = "^23.1.0"
cryptography = "^46.0.3"
Pillow = "^11.0.0"
pillow-heif = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
ruff = "^0.1.0"

[tool.poetry.scripts]
icloudbridge = "icloudbridge.cli.main:main_entry"
notes-ripper = "icloudbridge.scripts.rich_notes:main"
dev-server = "icloudbridge.scripts.dev_server:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"icloudbridge/cli/main.py" = ["UP007"]  # Typer requires Optional[X] syntax

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --cov=icloudbridge --cov-report=term-missing"

[tool.coverage.run]
omit = [
    "*/tests/*",
    "*/test_*.py",
]
